Bypass surgery on the user/firm backend code
Renamed UserProfile to AccountProvider.
It still extends model, so just a name change for the most part. Any
inquries for model definition will use User/FirmProfileDefinition
located in models/def/
Also cleaned up the interface to the provider to be able to handle
user and firm select and insert. (firm is untested)

Rewrote the post helper to handle the new definitions and added
support for array fields. Will clean this code up a bit later.

Still lots of verification and checking to do...

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>