Individual registration form ajax implementation
Processing of the registration and validation forms are working over
ajax. Errors are being trapped and sent back to an error box. The
process is using HTTP status code 202 to signify the form is validated
and a confirmation code has been sent to the registered email. Code
201 is used once the account is created after verification.

The data is being persisted through session based cookies for testing.
This needs to move to a server-side solution, or the cookies will
need encryption. Would be best to use a sql table.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>
Signed-off-by: Daniel Zukowski <daniel.zukowski@gmail.com>