Did more work with auth and started the cart
The authenticator is accessible from $this->mod_auth. It is loaded
through a hook so it doesn't need to be autoloaded. The authenticator
handles a Credentials object that is currently holding the auth status,
user id and user type. There is a user type for anonymous, normal, super
and child users.

Authenticator::grant was changed to take a user id and an array. The array
is any information that needs to be handled in the session. Currently
this is only the user type, it is also not extensible but this may change.

Setup the cart controller entry points to model a 4 step process that
branches at the start depending on if the user is a super user or not.
Super users need a different interface to manage attendees. Anon and
Normal users see a simple cart and can either login or checkout
respectively. Either case leads to a review page, then billing and
ends up with a final page that is printer friendly.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>