First stage of cart working for anonymous and normal users.
Anonymous users see Register/Login buttons instead of a checkout button.
Registration, if selected, forwards to login. The login looks up a referral
page in the session and loads back to the cart display. If the user authenticated
to a normal user, they see the page they just came from with a Checkout button.
Super users will see a cart with options to add attendees.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>