A few updates to auth and reg forms.
Authentication now fully tracks userType. Had to add the user type
to the db schema in order to get it back after login.

Update the firm profile and firm form to inject a hidden field with the
regtype selected in the first panel. The firm form needs it for
validation as its fields are dependant on its value. After the form
is submitted the userType gets exchanged to the user profile where it
goes for persistence.

As per http://github.com/Cluster444/nita/issues#issue/10 the validation
logic has gotten a bit wonky. I think its broken still with this
commit.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>