Update for cart display for super users.
Most of the UI logic is in place and seems to work ok. Tried
to break it in a few ways and it seems like it works. It doesn't
yet tie into the data backend, need that in place before Continue
can be hooked into the server-side.

Added the users accountId to the credentials.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>